Refactor : placeHeartService
@Transactional 제거 -> 필요 없음
중복코드 제거

+++ server/src/main/java/onde/there/place/service/PlaceHeartService.java
@@ -25,60 +25,67 @@ public class PlaceHeartService {
 	private final RedisServiceForPlaceHeart<Long> redisService;
 	private static final String PLACE_ID_KEY = "placeId";
 
-	@Transactional
 	public boolean heart(Long placeId, String memberId) {
+		boolean heart = true;
 		log.info("heart : 장소 좋아요 메소드 시작 (장소 아이디 : " + placeId + ") (맴버 아이디 : " + memberId + ")");
 		Place place = checkPlace(placeId);
 
 		Member member = checkMember(memberId);
 
-		if (placeHeartRepository.existsByPlaceIdAndMemberId(placeId, memberId)) {
-			throw new PlaceException(PlaceErrorCode.ALREADY_HEARTED);
-		}
+		checkPlaceHeart(place, member, heart);
 
-		placeHeartRepository.save(PlaceHeart.builder()
-			.place(place)
-			.member(member)
-			.build());
-
-		placeHeartUpdateRole(placeId, place, true);
+		placeHeartUpdateRole(placeId, place, heart);
 
 		log.info("heart : 장소 좋아요 메소드 완료 (장소 아이디 : " + placeId + ") (맴버 아이디 : " + memberId + ")");
 		return true;
 	}
 
-	@Transactional
 	public boolean unHeart(Long placeId, String memberId) {
+		boolean unHeart = false;
 		log.info(
 			"unHeart : 장소 좋아요 취소 메소드 시작 (장소 아이디 : " + placeId + ") (맴버 아이디 : " + memberId + ")");
 		Place place = checkPlace(placeId);
 
 		Member member = checkMember(memberId);
 
-		PlaceHeart placeHeart = placeHeartRepository.findByPlaceAndMember(place, member)
-			.orElseThrow(() -> new PlaceException(PlaceErrorCode.ALREADY_UN_HEARTED));
+		checkPlaceHeart(place, member, unHeart);
 
-		placeHeartRepository.delete(placeHeart);
+		placeHeartUpdateRole(placeId, place, unHeart);
 
-		placeHeartUpdateRole(placeId, place, false);
-
-		log.info("unHeart : 장소 좋아요 취소 메소드 완료 (장소 아이디 : " + placeId + ") (맴버 아이디 : " + memberId + ")");
+		log.info(
+			"unHeart : 장소 좋아요 취소 메소드 완료 (장소 아이디 : " + placeId + ") (맴버 아이디 : " + memberId + ")");
 		return true;
 	}
 
 	private Place checkPlace(Long placeId) {
-		Place place = placeRepository.findById(placeId)
+		return placeRepository.findById(placeId)
 			.orElseThrow(() -> new PlaceException(PlaceErrorCode.NOT_FOUND_PLACE));
-		return place;
 	}
 
 	private Member checkMember(String memberId) {
-		Member member = memberRepository.findById(memberId)
+		return memberRepository.findById(memberId)
 			.orElseThrow(() -> new PlaceException(PlaceErrorCode.NOT_FOUND_MEMBER));
-		return member;
 	}
 
-	private void placeHeartUpdateRole(Long placeId, Place place, boolean plusOrMinus) {
+	private void checkPlaceHeart(Place place, Member member, boolean heartOrUnHeart) {
+		if (heartOrUnHeart) {
+			if (placeHeartRepository.existsByPlaceIdAndMemberId(place.getId(), member.getId())) {
+				throw new PlaceException(PlaceErrorCode.ALREADY_HEARTED);
+			}
+
+			placeHeartRepository.save(PlaceHeart.builder()
+				.place(place)
+				.member(member)
+				.build());
+		} else {
+			PlaceHeart placeHeart = placeHeartRepository.findByPlaceAndMember(place, member)
+				.orElseThrow(() -> new PlaceException(PlaceErrorCode.ALREADY_UN_HEARTED));
+
+			placeHeartRepository.delete(placeHeart);
+		}
+	}
+
+	private void placeHeartUpdateRole(Long placeId, Place place, boolean heartOrUnHeart) {
 		log.info("placeHeartUpdateRole : 장소 좋아요 갯수 업데이트 메소드 시작! (장소 아이디 : " + placeId + ")");
 		if (place.getPlaceHeartCount() >= 1000) {
 			addSchedule(placeId);
@@ -86,7 +93,7 @@ public class PlaceHeartService {
 				"좋아요 갯수 1000개 이상 -> placeHeartUpdateRole : 장소 좋아요 갯수 업데이트 메소드 취소 -> 스케줄링으로 저장 완료! (장소 아이디 : "
 					+ placeId + ")");
 		} else {
-			place.setPlaceHeartCount(place.getPlaceHeartCount() + (plusOrMinus ? 1 : -1));
+			place.setPlaceHeartCount(place.getPlaceHeartCount() + (heartOrUnHeart ? 1 : -1));
 			placeRepository.save(place);
 			log.info("좋아요 갯수 1000개 미만 -> placeHeartUpdateRole : 장소 좋아요 갯수 업데이트 메소드 완료! (장소 아이디 : "
 				+ placeId + ")");
